requirements.txt
alembic==0.7.4
flasgger==0.5.13
Flask==1.0.2
Flask-Migrate==1.2.0
Flask-RESTful-TheodoFork==0.3.7
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
itsdangerous==0.24
Jinja2==2.10
psycopg2==2.7.1
pytz==2014.10
requests==2.20.1
setproctitle==1.1.9
six==1.9.0
SQLAlchemy==1.1.3
Werkzeug==0.14.1
python-dotenv
